# gbs-pos-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Point of Sale REST API for retail, F\u0026B, and fuel station operations",
        "title": "GBS POS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GBS Support",
            "email": "support@gbs.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/v1/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get orders with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store type filter",
                        "name": "storeType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment method filter",
                        "name": "paymentMethod",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Terminal ID filter",
                        "name": "terminalId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Start date Unix timestamp",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "End date Unix timestamp",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter voided orders",
                        "name": "isVoided",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter settled orders",
                        "name": "isSettled",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create order",
                "parameters": [
                    {
                        "description": "Order data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Idempotent response (order already exists"
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/v1/orders/bulk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sync multiple orders at once (offline support)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Bulk sync orders",
                "parameters": [
                    {
                        "description": "Orders to sync",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkSyncOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/v1/orders/settle": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Settle/close orders for a terminal (ADMIN only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Settle orders",
                "parameters": [
                    {
                        "description": "Settlement data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SettleOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/v1/orders/unsettled/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get summary of unsettled orders for terminal",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get unsettled orders summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Store type filter",
                        "name": "storeType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Terminal ID filter",
                        "name": "terminalId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/v1/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a single order by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/v1/orders/{id}/void": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Void/cancel an order (ADMIN only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Void order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Void reason",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VoidOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
...
        "/v1/sync/orders": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sync multiple orders at once (offline support)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Bulk sync orders",
                "parameters": [
                    {
                        "description": "Orders to sync",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkSyncOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.BulkSyncOrderRequest": {
            "type": "object",
            "required": [
                "orders"
            ],
            "properties": {
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Order"
                    }
                },
                "terminalId": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateOrderRequest": {
            "type": "object",
            "required": [
                "id",
                "items",
                "paymentMethod",
                "subtotal",
                "tax",
                "timestamp",
                "total"
            ],
            "properties": {
                "acqMid": {
                    "type": "string"
                },
                "acqTid": {
                    "type": "string"
                },
                "approvalCode": {
                    "type": "string"
                },
                "bankName": {
                    "type": "string"
                },
                "cashReceived": {
                    "type": "number"
                },
                "changeAmount": {
                    "type": "number"
                },
                "customerId": {
                    "type": "integer"
                },
                "customerName": {
                    "type": "string"
                },
                "customerPhone": {
                    "type": "string"
                },
                "discountType": {
                    "type": "string"
                },
                "discountValue": {
                    "type": "number"
                },
                "entryMode": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "productId",
                            "productName",
                            "productPrice",
                            "qty",
                            "subtotal"
                        ],
                        "properties": {
                            "productId": {
                                "type": "integer"
                            },
                            "productName": {
                                "type": "string"
                            },
                            "productPrice": {
                                "type": "number"
                            },
                            "qty": {
                                "type": "integer"
                            },
                            "sku": {
                                "type": "string"
                            },
                            "subtotal": {
                                "type": "number"
                            },
                            "variantId": {
                                "type": "integer"
                            },
                            "variantName": {
                                "type": "string"
                            }
                        }
                    }
                },
                "maskedAccount": {
                    "type": "string"
                },
                "paymentMethod": {
                    "type": "string"
                },
                "posMessageId": {
                    "type": "string"
                },
                "storeType": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                },
                "tax": {
                    "type": "number"
                },
                "terminalId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                },
                "transactionId": {
                    "type": "string"
                }
            }
        },
...
        "dto.SettleOrderRequest": {
            "type": "object",
            "required": [
                "settlementId",
                "timestamp"
            ],
            "properties": {
                "settlementId": {
                    "type": "string"
                },
                "storeType": {
                    "type": "string"
                },
                "terminalId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "dto.VoidOrderRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "model.Order": {
            "type": "object",
            "properties": {
                "acqMid": {
                    "type": "string"
                },
                "acqTid": {
                    "type": "string"
                },
                "approvalCode": {
                    "type": "string"
                },
                "bankName": {
                    "type": "string"
                },
                "cashReceived": {
                    "type": "number"
                },
                "changeAmount": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "customerId": {
                    "type": "integer"
                },
                "customerName": {
                    "type": "string"
                },
                "customerPhone": {
                    "type": "string"
                },
                "discountAmount": {
                    "type": "number"
                },
                "discountType": {
                    "type": "string"
                },
                "discountValue": {
                    "type": "number"
                },
                "entryMode": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isSettled": {
                    "type": "boolean"
                },
                "isVoided": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrderItem"
                    }
                },
                "loyaltyPointsEarned": {
                    "type": "integer"
                },
                "maskedAccount": {
                    "type": "string"
                },
                "paymentMethod": {
                    "type": "string"
                },
                "posMessageId": {
                    "type": "string"
                },
                "storeType": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                },
                "tax": {
                    "type": "number"
                },
                "terminalId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "total": {
                    "type": "number"
                },
                "transactionId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "voidReason": {
                    "type": "string"
                },
                "voidedAt": {
                    "type": "string"
                },
                "voidedBy": {
                    "type": "string"
                }
            }
        },
        "model.OrderItem": {
            "type": "object",
            "properties": {
                "productId": {
                    "type": "integer"
                },
                "productName": {
                    "type": "string"
                },
                "productPrice": {
                    "type": "number"
                },
                "qty": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                },
                "variantId": {
                    "type": "integer"
                },
                "variantName": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. For Keycloak, obtain token from {KEYCLOAK_BASE_URL}/realms/{KEYCLOAK_REALM}/protocol/openid-connect/token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}